installs

+++ dotfiles/scripts_install/install-mutt.sh
@@ -23,10 +23,11 @@ fi
 
 if [ ! -e $TARGET ]; then
 
-  TAR_NAME=$NAME_VERSION.tar.bz2
+  TAR_NAME="${NAME_VERSION}a.tar.bz2"
   TAR=$ROOT/$TAR_NAME
+  RET=1
 
-  wget "http://www.jedsoft.org/releases/slang/$TAR_NAME" -O $TAR
+  wget "http://www.jedsoft.org/releases/$NAME/$TAR_NAME" -O $TAR
   if [ $? -eq 0 ];then
     mkdir $TARGET
     tar -xf $TAR -C $TARGET --strip-components=1
@@ -46,13 +47,14 @@ if [ ! -e $TARGET ]; then
           fi
           sudo make install
           if [ $? -eq 0 ];then
+            RET=0
             ln -s $TARGET $LATEST
-            echo "SLANG OK"
+            echo "$NAME OK"
           fi
         fi
       fi
     fi
-    if [ ! $? -eq 0 ];then
+    if [ ! $RET -eq 0 ];then
       rm -rf $TARGET
     fi
     rm $TAR
@@ -77,8 +79,9 @@ if [ ! -e $TARGET ]; then
 
   TAR_NAME=$NAME_VERSION.tar.gz
   TAR=$ROOT/$TAR_NAME
+  RET=1
 
-  wget "http://ftp.gnu.org/pub/gnu/libiconv/$TAR_NAME" -O $TAR
+  wget "http://ftp.gnu.org/pub/gnu/$NAME/$TAR_NAME" -O $TAR
   if [ $? -eq 0 ];then
     mkdir $TARGET
     tar -xzf $TAR -C $TARGET --strip-components=1
@@ -98,13 +101,71 @@ if [ ! -e $TARGET ]; then
           fi
           sudo make install
           if [ $? -eq 0 ];then
+            RET=0
             ln -s $TARGET $LATEST
-            echo "LIBICONV OK"
+            echo "$NAME OK"
           fi
         fi
       fi
     fi
-    if [ ! $? -eq 0 ];then
+    if [ ! $RET -eq 0 ];then
+      rm -rf $TARGET
+    fi
+    rm $TAR
+  fi
+fi
+# ===================================================
+# ===================================================
+# openssl
+# ===================================================
+# ===================================================
+NAME=openssl
+VERSION="1.0.2j"
+NAME_VERSION=$NAME-$VERSION
+ROOT=$SOURCES_ROOT/$NAME
+TARGET=$ROOT/$NAME_VERSION
+LATEST=$ROOT/$NAME-latest
+if [ ! -e $ROOT ];then
+  mkdir $ROOT
+fi
+
+if [ ! -e $TARGET ]; then
+
+  TAR_NAME=$NAME_VERSION.tar.gz
+  TAR=$ROOT/$TAR_NAME
+  RET=1
+  wget "https://openssl.org/source/$TAR_NAME" -O $TAR
+  if [ $? -eq 0 ];then
+    mkdir $TARGET
+    tar -xzf $TAR -C $TARGET --strip-components=1
+    if [ $? -eq 0 ];then 
+      cd $TARGET
+      ./config --prefix=/usr/local --openssldir=/etc/ssl --libdir=lib shared zlib-dynamic
+      if [ $? -eq 0 ]; then
+        make depend
+        if [ $? -eq 0 ]; then
+          make
+          if [ $? -eq 0 ]; then
+            CURRENT=`pwd`
+            if [ -e $LATEST ];then
+              make MANDIR=/usr/share/man MANSUFFIX=ssl install && install -dv -m755 /usr/share/doc/openssl-1.0.2j  && cp -vfr doc/*     /usr/share/doc/openssl-1.0.2j
+              # uninstall previous
+              cd $LATEST
+              sudo make uninstall
+              cd $CURRENT
+              rm -rf $LATEST
+            fi
+            sudo make install
+            if [ $? -eq 0 ];then
+              RET=0
+              ln -s $TARGET $LATEST
+              echo "$NAME OK"
+            fi
+          fi
+        fi
+      fi
+    fi
+    if [ ! $RET -eq 0 ];then
       rm -rf $TARGET
     fi
     rm $TAR
@@ -186,8 +247,9 @@ if [ ! -e $TARGET ]; then
 
   TAR_NAME=$NAME_VERSION.tar.gz
   TAR=$ROOT/$TAR_NAME
+  RET=1
 
-  wget "ftp://ftp.gnu.org/gnu/gdbm/$TAR_NAME" -O $TAR
+  wget "ftp://ftp.gnu.org/gnu/$NAME/$TAR_NAME" -O $TAR
   if [ $? -eq 0 ];then
     mkdir $TARGET
     tar -xzf $TAR -C $TARGET --strip-components=1
@@ -207,13 +269,14 @@ if [ ! -e $TARGET ]; then
           fi
           sudo make install
           if [ $? -eq 0 ];then
+            RET=0
             ln -s $TARGET $LATEST
-            echo "gdbm OK"
+            echo "$NAME OK"
           fi
         fi
       fi
     fi
-    if [ ! $? -eq 0 ];then
+    if [ ! $RET -eq 0 ];then
       rm -rf $TARGET
     fi
     rm $TAR
@@ -241,22 +304,28 @@ if [ ! -e $TARGET ]; then
 
   PATCH_NAME=$NAME_VERSION-fixes-3.patch
   PATCH_FILE=$ROOT/$PATCH_NAME
+  RET=1
 
-  wget "ftp://ftp.cyrusimap.org/cyrus-sasl/$TAR_NAME" -O $TAR
+  wget "ftp://ftp.cyrusimap.org/$NAME/$TAR_NAME" -O $TAR
   if [ $? -eq 0 ];then
     wget "http://www.linuxfromscratch.org/patches/blfs/svn/$PATCH_NAME" -O $PATCH_FILE
     if [ $? -eq 0 ];then
       mkdir $TARGET
+      echo "------------------tar"
       tar -xzf $TAR -C $TARGET --strip-components=1
       if [ $? -eq 0 ];then 
         cd $TARGET
+        echo "------------------patch"
         patch -Np1 -i $PATCH_FILE
         if [ $? -eq 0 ]; then
-          autoconf -fi
+          echo "------------------autoreconf"
+          autoreconf -fi
           if [ $? -eq 0 ]; then
             # --sysconfdir=/etc --enable-auth-sasldb --with-dbpath=/var/lib/sasl/sasldb2 --with-saslauthd=/var/run/saslauthd
+            echo "------------------configure"
             ./configure --prefix=/usr/local
             if [ $? -eq 0 ]; then
+              echo "------------------make"
               make
               if [ $? -eq 0 ]; then
                 CURRENT=`pwd`
@@ -267,10 +336,13 @@ if [ ! -e $TARGET ]; then
                   cd $CURRENT
                   rm -rf $LATEST
                 fi
+                echo "------------------make install"
                 sudo make install
                 if [ $? -eq 0 ];then
+                  RET=0
                   ln -s $TARGET $LATEST
                   sudo ln -s /usr/local/lib/sasl2 /usr/lib/sasl2
+                  sudo ln -s /usr/local/include/sasl /usr/include/sasl
                   echo "SASL OK"
                 fi
               fi
@@ -279,7 +351,7 @@ if [ ! -e $TARGET ]; then
         fi
       fi
     fi
-    if [ ! $? -eq 0 ];then
+    if [ ! $RET -eq 0 ];then
       rm -rf $TARGET
     fi
     rm $TAR
@@ -292,9 +364,10 @@ fi
 # mutt
 # ===================================================
 # ===================================================
+NAME=mutt
 MUTT_VERSION=1.7.2
-MUTT=mutt-$MUTT_VERSION
-MUTT_ROOT=$SOURCES_ROOT/mutt
+MUTT=$NAME-$MUTT_VERSION
+MUTT_ROOT=$SOURCES_ROOT/$NAME
 TARGET=$MUTT_ROOT/$MUTT
 if [ ! -e $MUTT_ROOT ];then
   mkdir $MUTT_ROOT
@@ -310,13 +383,15 @@ sudo chgrp -v mail $VAR_MAIL
 if [ ! -e $TARGET ]; then
   MUTT_TAR_NAME=$MUTT.tar.gz
   MUTT_TAR=$MUTT_ROOT/$MUTT_TAR_NAME
+  RET=1
 
   cd $MUTT_ROOT
-  wget "ftp://ftp.mutt.org/pub/mutt/$MUTT_TAR_NAME" -O $MUTT_TAR
+  wget "ftp://ftp.mutt.org/pub/$NAME/$MUTT_TAR_NAME" -O $MUTT_TAR
   if [ $? -eq 0 ];then
     mkdir $TARGET
     tar -xzf $MUTT_TAR -C $TARGET --strip-components=1
     if [ $? -eq 0 ];then
+      cd $TARGET
       ./configure --prefix=/usr --with-docdir=/usr/share/doc/$MUTT --with-openssl --with-ssl --with-slang --with-sasl --sysconfdir=/etc --enable-external-dotlock --enable-pop --enable-imap --enable-smtp --enable-hcache --enable-sidebar
       if [ $? -eq 0 ]; then
         make
@@ -331,6 +406,7 @@ if [ ! -e $TARGET ]; then
           fi
           sudo make install
           if [ $? -eq 0 ];then
+            RET=0
             ln -s $TARGET $LATEST
             echo "MUTT OK"
           fi
@@ -338,7 +414,7 @@ if [ ! -e $TARGET ]; then
       fi
     fi
   fi
-  if [ ! $? -eq 0 ];then
+  if [ ! $RET -eq 0 ];then
     rm -rf $TARGET
   fi
   rm $MUTT_TAR
